requirements.txt
appdirs==1.4.0
coreapi==2.1.1
Django==1.10.5
django-cors-headers==2.0.1
django-filter==1.0.1
django-model-utils==2.6.1
django-notifications-hq==1.2
django-rest-swagger==2.1.1
djangorestframework==3.5.3
itypes==1.1.0
jsonfield==1.0.3
Markdown==2.6.7
olefile==0.44
openapi-codec==1.2.1
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
rest-framework-generic-relations==1.1.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
uWSGI==2.0.14
